+++ minijinja/src/lib.rs
@@ -101,6 +101,7 @@
 //!   of things like callbacks however are not changing which means code that uses
 //!   MiniJinja still needs to be threadsafe.
 //! - `json`: When enabled the `tojson` filter is added as builtin filter.
+//! - `urlencode`: When enabled the `urlencode` filter is added as builtin filter.
 #![allow(clippy::cognitive_complexity)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 mod key;